Move node_config.ts
